test(admission): prove the meter is live on the responder, and satisfy the gates (#269)
Adds the WIRING assertion the unit tests cannot make: `admission.rs`'s tests prove
the meter behaves, not that anything calls it, and a meter nothing calls is the
defect #269 describes. The new test drives `NodeResponder::handle_json_rpc` and
asserts an unauthenticated session is refused with -32000 rather than the
allowlist's -32601 -- which is how we know admission ran FIRST -- with an
authenticated control proving the responder is not simply refusing everybody.

`node_responder_returns_method_not_found_for_management_methods` now passes a
64-hex conn_key. That is the shape every production session supplies, since both
listeners derive it from the verified client leaf and no production path reaches
the responder caller-less. Without it the test would answer -32000 and stop
exercising the allowlist at all; with it, the property under test is unchanged.

Also drops an unused accessor and two redundant u32 conversions for clippy
-D warnings.

+++ crates/dig-node-core/src/seams/dig_peer/conduct.rs
@@ -68,7 +68,10 @@ impl ConductState {
     /// identity would be *better* than an honest long-lived one and the ranking would reward churn.
     pub(crate) fn observe(&self, peer: RoutedPeer, evidence: ConductEvidence, now_ticks: u64) {
         let mut records = self.lock();
-        let record = records.get(&peer).copied().unwrap_or_else(ConductRecord::neutral);
+        let record = records
+            .get(&peer)
+            .copied()
+            .unwrap_or_else(ConductRecord::neutral);
         records.insert(peer, dig_sex::observe(record, evidence, now_ticks));
     }
 
@@ -214,7 +217,7 @@ mod tests {
     fn sustained_non_performance_never_silences_a_peer_completely() {
         let conduct = ConductState::new();
         let distressed = peer(0xd4);
-        for _ in 0..(u32::from(dig_sex::NON_PERFORMANCE_CEILING) * 10) {
+        for _ in 0..(dig_sex::NON_PERFORMANCE_CEILING * 10) {
             conduct.observe(distressed, ConductEvidence::NonPerformance, 0);
         }
 
@@ -280,7 +283,7 @@ mod tests {
 
         conduct.observe(liar, ConductEvidence::ProvenLie, 0);
         // Far past the ceiling: as much non-performance as an attacker could ever manufacture.
-        for _ in 0..(u32::from(dig_sex::NON_PERFORMANCE_CEILING) * 10) {
+        for _ in 0..(dig_sex::NON_PERFORMANCE_CEILING * 10) {
             conduct.observe(slow, ConductEvidence::NonPerformance, 0);
         }
